+++ spec/workspace_spec.rb
@@ -8,13 +8,15 @@ describe "Xcodeproj::Workspace" do
 
     it "accepts new projects" do
       @workspace << 'Framework.xcodeproj'
-      @workspace.projpaths.should.include 'Framework.xcodeproj'
+      @workspace.file_references.should.include 'Framework.xcodeproj'
     end
   end
 
   describe "converted to XML" do
     before do
-      @workspace = Xcodeproj::Workspace.new('Pods/Pods.xcodeproj', 'App.xcodeproj')
+      pods_project_file_reference = Xcodeproj::Workspace::FileReference.new('Pods/Pods.xcodeproj')
+      project_file_reference = Xcodeproj::Workspace::FileReference.new('App.xcodeproj')
+      @workspace = Xcodeproj::Workspace.new(pods_project_file_reference, project_file_reference)
       @doc = REXML::Document.new(@workspace.to_s)
     end
 
@@ -35,9 +37,9 @@ describe "Xcodeproj::Workspace" do
     end
 
     it "contains all of the projects in the workspace" do
-      @workspace.projpaths.should.include "libPusher.xcodeproj"
-      @workspace.projpaths.should.include "libPusher-OSX/libPusher-OSX.xcodeproj"
-      @workspace.projpaths.should.include "Pods/Pods.xcodeproj"
+      @workspace.file_references.should.include Xcodeproj::Workspace::FileReference.new("libPusher.xcodeproj")
+      @workspace.file_references.should.include Xcodeproj::Workspace::FileReference.new("libPusher-OSX/libPusher-OSX.xcodeproj")
+      @workspace.file_references.should.include Xcodeproj::Workspace::FileReference.new("Pods/Pods.xcodeproj")
     end
   end
 
@@ -47,36 +49,36 @@ describe "Xcodeproj::Workspace" do
     end
 
     it "contains no projects" do
-      @workspace.projpaths.should.be.empty
+      @workspace.file_references.should.be.empty
     end
   end
-  
+
   describe "load schemes for all projects from workspace file" do
     before do
       @workspace = Xcodeproj::Workspace.new_from_xcworkspace(fixture_path("SharedSchemes/SharedSchemes.xcworkspace"))
     end
-    
+
     it "returns data type should be hash" do
       @workspace.schemes.should.instance_of Hash
     end
-    
-    it "schemes count should be greater or equal to projpaths count" do
-      @workspace.schemes.count.should >= @workspace.projpaths.count
+
+    it "schemes count should be greater or equal to file_references count" do
+      @workspace.schemes.count.should >= @workspace.file_references.count
     end
-    
+
     it "contains only test data schemes" do
       @workspace.schemes.keys.sort.should == ['Pods', 'SharedSchemes', 'SharedSchemesForTest']
     end
   end
-  
+
   describe "built from a workspace file with XML entities in a project path" do
     before do
       @workspace = Xcodeproj::Workspace.new_from_xcworkspace(fixture_path("Otto's Remote.xcworkspace"))
     end
 
     it "contains all of the projects in the workspace" do
-      @workspace.projpaths.should.include "Otto's Remote.xcodeproj"
-      @workspace.projpaths.should.include "Pods/Pods.xcodeproj"
+      @workspace.file_references.should.include Xcodeproj::Workspace::FileReference.new("Otto's Remote.xcodeproj")
+      @workspace.file_references.should.include Xcodeproj::Workspace::FileReference.new("Pods/Pods.xcodeproj")
     end
   end
 end